Refactored addCourse
