+++ app/addCourse/page.jsx
@@ -1,13 +1,11 @@
 'use client';
 import { useState, useEffect } from 'react';
-import { collection, doc, getDocs, getDoc, setDoc } from 'firebase/firestore';
-import db from '../lib/firebase';
 import { onAuthStateChanged } from 'firebase/auth';
 import { auth } from '../lib/firebase';
-import { query, where } from "firebase/firestore";
-import { set } from 'firebase/database';
 import AdminSidebar from '../components/AdminSidebar';
-import { createUser } from '../models/User';
+import { isCourseCodeValid, courseAlreadyExists, addCourse } from '../models/Course';
+import { createUser, fetchAllTeachers, getTeacherDoc, addTeacherRegisteredCourse } from '../models/User';
+import AdminAddCourseView from '../views/AdminAddCourseView';
 
 const AddCoursePage = () => {
     const [user,setUser] = useState(null);
@@ -16,29 +14,7 @@ const AddCoursePage = () => {
     const [description, setDescription] = useState('');
     const [teachers, setTeachers] = useState(['Fetching all teachers...']);
     const [selectedTeacher, setSelectedTeacher] = useState('');
-    const [loading, setLoading] = useState(true);
-
-    const fetchTeachers = async () => {
-        try {
-            const teachersRef = collection(db, 'teachers');
-            const snapshot = await getDocs(teachersRef);
-            const teachersData = snapshot.docs.map(doc => ({ id: doc.id, ...doc.data() }));
-            setTeachers(teachersData);
-        } catch (error) {
-            console.error('Error fetching teachers:', error);
-        }
-    };
-
-    const isCourseCodeValid = (courseCode) => {
-        const regex = /^[A-Z]{4}\d{3}$/;
-        return regex.test(courseCode);
-    }
-
-    const courseAlreadyExists = async (courseCode) => {
-        const courseRef = doc(db, 'courses', courseCode);
-        const courseSnapshot = await getDoc(courseRef);
-        return courseSnapshot.exists();
-    }
+    const [loading, setLoading] = useState(true);       // TODO: use loading state to show a spinner while fetching data?
 
     const handleInputChange = (e) => {
         const { name, value } = e.target;
@@ -86,20 +62,14 @@ const AddCoursePage = () => {
         };
     
         try{
-            const coursesRef = collection(db, ' courses');
-            await setDoc(doc(db, "courses", courseCode), newCourseData);
+            await addCourse(courseCode, newCourseData);
+            console.log("Course added to courses collection");
+            
+            const teacherDoc = await getTeacherDoc(selectedTeacher);
+            console.log("Got teacher snapshot");
 
-            console.log("Course added to courses collection")
-
-            const teacherRef = doc(db, 'teachers', selectedTeacher);
-            const teacherDoc = await getDoc(teacherRef);
-        
-            console.log("Got teacher snapshot")
-
-            const registeredCoursesRef = collection( teacherDoc.ref, 'registeredCourses');
-            await setDoc(doc(registeredCoursesRef, courseCode), newCourseData); 
-
-            console.log("Course added to registered courses")
+            await addTeacherRegisteredCourse(teacherDoc.ref, courseCode, newCourseData);
+            console.log("Course added to teacher's registered courses");
 
             setCourseCode('');
             setCourseName('');
@@ -109,16 +79,19 @@ const AddCoursePage = () => {
         } catch(error){
             console.log("Error adding course:", error);
         }
-
     };
 
     useEffect(() => {
         const unsubscribe = onAuthStateChanged(auth, async (user) => {
             if(auth.currentUser){
                 if(auth.currentUser){
+                    // User is signed in
                     const user = await createUser(auth.currentUser.uid, "admin");
                     setUser(user);
-                    fetchTeachers();
+
+                    const teachers = await fetchAllTeachers();
+                    setTeachers(teachers);
+
                     setTimeout(() => {
                         setLoading(false);
                     }, 200);
@@ -136,70 +109,16 @@ const AddCoursePage = () => {
     return (
         <div className="flex flex-col justify-center items-center h-screen bg-gray-200">
             <AdminSidebar data-testid="adminSidebar-component" userName={ user?.firstName } />
-            <h1 className='text-black mb-32 font-bold text-3xl'>Add Course</h1>
-            <form className="bg-white shadow-md rounded px-8 pt-6 pb-8 mb-4" onSubmit={handleSubmit}>
-                <div className="mb-4">
-                    <label className="block text-gray-700 text-sm font-bold mb-2" htmlFor="courseCode">
-                        Course Code:
-                    </label>
-                    <input
-                        className="shadow appearance-none border rounded w-full py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:shadow-outline"
-                        id="courseCode"
-                        type="text"
-                        name="courseCode"
-                        value={courseCode}
-                        onChange={handleInputChange}
-                    />
-                </div>
-                <div className="mb-4">
-                    <label className="block text-gray-700 text-sm font-bold mb-2" htmlFor="courseName">
-                        Course Name:
-                    </label>
-                    <input
-                        className="shadow appearance-none border rounded w-full py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:shadow-outline"
-                        id="courseName"
-                        type="text"
-                        name="courseName"
-                        value={courseName}
-                        onChange={handleInputChange}
-                    />
-                </div>
-                <div className="mb-4">
-                    <label className="block text-gray-700 text-sm font-bold mb-2" htmlFor="description">
-                        Description:
-                    </label>
-                    <textarea
-                        className="shadow appearance-none border rounded w-full py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:shadow-outline h-40"
-                        id="description"
-                        type="text"
-                        name="description"
-                        value={description}
-                        onChange={handleInputChange}
-                    />
-                </div>
-                <div className="mb-4">
-                    <label className="block text-gray-700 text-sm font-bold mb-2" htmlFor="teacher">
-                        Teacher:
-                    </label>
-                    <select
-                        className="shadow appearance-none border rounded w-full py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:shadow-outline"
-                        id="teacher"
-                        name="teacher"
-                        value={selectedTeacher}
-                        onChange={handleInputChange}
-                    >
-                        <option value="">Select Teacher</option>
-                        {teachers.map((teacher) => (
-                            <option key={teacher.id} value={teacher.id}>{teacher.firstName} {teacher.lastName}</option>
-                        ))}
-                    </select>
-                </div>
-                <div className="flex items-center justify-between">
-                    <button className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded focus:outline-none focus:shadow-outline" type="submit">
-                        Submit
-                    </button>
-                </div>
-            </form>
+            <AdminAddCourseView 
+                loading={loading}
+                handleInputChange={handleInputChange} 
+                handleSubmit={handleSubmit} 
+                courseCode={courseCode} 
+                courseName={courseName} 
+                description={description} 
+                selectedTeacher={selectedTeacher} 
+                teachers={teachers} 
+            />
         </div>
     );
 };
