--- app/components/FetchCourseData.mock.js
@@ -1,5 +1,0 @@
-export const fetchCourseInfo = () => Promise.resolve([
-    { id: '1', courseCode: 'course1', courseName: 'Course 1', imageUrl: 'https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQ_gaZXWjn_qJVUXTVnHnGIPRka3psRSJgShg&usqp=CAU' },
-    { id: '2', courseCode: 'course2', courseName: 'Course 2', imageUrl: 'https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQ_gaZXWjn_qJVUXTVnHnGIPRka3psRSJgShg&usqp=CAU' },
-    // Add more dummy courses as needed
-]);
