--- app/components/FetchCourseDataForTest.js
@@ -1,4 +1,0 @@
-import FetchCourseData from './FetchCourseData';
-
-export const fetchStudentCourseInfo = FetchCourseData.fetchStudentCourseInfo;
-export const fetchTeacherCourseInfo = FetchCourseData.fetchTeacherCourseInfo;
