+++ app/views/AdminAddCourseView.jsx
@@ -1,0 +1,71 @@
+// TODO: use loading? seems to work without it 
+export default function AdminAddCourseView({ handleInputChange, handleSubmit, courseCode, courseName, description, selectedTeacher, teachers, loading }) {
+    return (
+        <div>
+            <h1 className='text-black mb-32 font-bold text-3xl'>Add Course</h1>
+            <form className="bg-white shadow-md rounded px-8 pt-6 pb-8 mb-4" onSubmit={handleSubmit}>
+                <div className="mb-4">
+                    <label className="block text-gray-700 text-sm font-bold mb-2" htmlFor="courseCode">
+                        Course Code:
+                    </label>
+                    <input
+                        className="shadow appearance-none border rounded w-full py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:shadow-outline"
+                        id="courseCode"
+                        type="text"
+                        name="courseCode"
+                        value={courseCode}
+                        onChange={handleInputChange}
+                    />
+                </div>
+                <div className="mb-4">
+                    <label className="block text-gray-700 text-sm font-bold mb-2" htmlFor="courseName">
+                        Course Name:
+                    </label>
+                    <input
+                        className="shadow appearance-none border rounded w-full py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:shadow-outline"
+                        id="courseName"
+                        type="text"
+                        name="courseName"
+                        value={courseName}
+                        onChange={handleInputChange}
+                    />
+                </div>
+                <div className="mb-4">
+                    <label className="block text-gray-700 text-sm font-bold mb-2" htmlFor="description">
+                        Description:
+                    </label>
+                    <textarea
+                        className="shadow appearance-none border rounded w-full py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:shadow-outline h-40"
+                        id="description"
+                        type="text"
+                        name="description"
+                        value={description}
+                        onChange={handleInputChange}
+                    />
+                </div>
+                <div className="mb-4">
+                    <label className="block text-gray-700 text-sm font-bold mb-2" htmlFor="teacher">
+                        Teacher:
+                    </label>
+                    <select
+                        className="shadow appearance-none border rounded w-full py-2 px-3 text-gray-700 leading-tight focus:outline-none focus:shadow-outline"
+                        id="teacher"
+                        name="teacher"
+                        value={selectedTeacher}
+                        onChange={handleInputChange}
+                    >
+                        <option value="">Select Teacher</option>
+                        {teachers.map((teacher) => (
+                            <option key={teacher.id} value={teacher.id}>{teacher.firstName} {teacher.lastName}</option>
+                        ))}
+                    </select>
+                </div>
+                <div className="flex items-center justify-between">
+                    <button className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded focus:outline-none focus:shadow-outline" type="submit">
+                        Submit
+                    </button>
+                </div>
+            </form>
+        </div>
+    );
+}
